added UserList to Project Page

+++ src/global/components/ItemList/index.tsx
@@ -19,24 +19,28 @@ interface ItemListProps {
   onAddButtonClick: Function | any;
   addMessage: string;
   isLoading: boolean | any;
+  className?: string;
 }
 export const ItemList: React.FC<ItemListProps> = ({
   title,
   children,
   onAddButtonClick,
   addMessage,
-  isLoading
+  isLoading,
+  className
 }) => (
-  <StyledAside>
+  <StyledAside className={className}>
     <StyledTitle>{title}</StyledTitle>
     <StyledUl>
       {children}
       {isLoading && <StyledLoading />}
     </StyledUl>
-    <StyledAddSection onClick={onAddButtonClick}>
-      <AddButton position="static" onClick={onAddButtonClick} />
-      <p>{addMessage}</p>
-    </StyledAddSection>
+    {onAddButtonClick && (
+      <StyledAddSection onClick={onAddButtonClick}>
+        <AddButton position="static" onClick={onAddButtonClick} />
+        <p>{addMessage}</p>
+      </StyledAddSection>
+    )}
   </StyledAside>
 );
 
@@ -44,25 +48,34 @@ interface ItemProps {
   image?: string;
   name: string;
   direction?: string;
-  onDelete: Function;
+  onDelete?: Function;
+  onClick?: Function;
 }
 export const Item: React.FC<ItemProps> = ({
   image,
   name,
   direction,
-  onDelete
+  onDelete,
+  onClick
 }) => {
+  const [showDelete, setShowDelete] = React.useState(false);
+  const handleMouseHover = () => setShowDelete(true);
+  const handleMouseOut = () => setShowDelete(false);
   const handleDelete = (e: Event) => {
     //TODO: for some reason without this the router sends to main page. ??
     e.preventDefault();
-    onDelete();
+    onDelete && onDelete();
+  };
+  const handleClick = (e: Event) => {
+    e.preventDefault();
+    onClick && onClick();
   };
   return (
-    <StyledLi>
-      <StyledLink to={direction || ""}>
-        <StyledPicture>{image && <img src="" alt={image} />}</StyledPicture>
+    <StyledLi onMouseOver={handleMouseHover} onMouseLeave={handleMouseOut}>
+      <StyledLink to={direction || ""} onClick={handleClick}>
+        <StyledPicture>{image && <img src={image} alt="" />}</StyledPicture>
         <p>{name}</p>
-        <StyledDeleteIcon onClick={handleDelete} />
+        {onDelete && showDelete && <StyledDeleteIcon onClick={handleDelete} />}
       </StyledLink>
     </StyledLi>
   );
